+++ Models/Product.cs
@@ -7,7 +7,7 @@ namespace api.Models {
     public class Product{
         public int ProductId { get; set; }
         public string Name { get; set; } = string.Empty;
-        public int CIP { get; set; }
+        public long CIP { get; set; }
         public string DCI { get; set; } = string.Empty;
         public string Dosage { get; set; } = string.Empty;
         public bool FlagIsDelete { get; set; }
